+++ src/Engines/CollectionEngine.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace Yajra\Datatables\Engines;
+namespace Boparaiamrit\Datatables\Engines;
 
 use Closure;
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
-use Yajra\Datatables\Request;
+use Boparaiamrit\Datatables\Request;
 
 /**
  * Class CollectionEngine.
  *
- * @package Yajra\Datatables\Engines
+ * @package Boparaiamrit\Datatables\Engines
  * @author  Arjay Angeles <aqangeles@gmail.com>
  */
 class CollectionEngine extends BaseEngine
@@ -35,7 +35,7 @@ class CollectionEngine extends BaseEngine
      * CollectionEngine constructor.
      *
      * @param \Illuminate\Support\Collection $collection
-     * @param \Yajra\Datatables\Request $request
+     * @param \Boparaiamrit\Datatables\Request $request
      */
     public function __construct(Collection $collection, Request $request)
     {