+++ src/Html/Builder.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Yajra\Datatables\Html;
+namespace Boparaiamrit\Datatables\Html;
 
 use Collective\Html\FormBuilder;
 use Collective\Html\HtmlBuilder;
@@ -14,7 +14,7 @@ use Illuminate\Support\Str;
 /**
  * Class Builder.
  *
- * @package Yajra\Datatables\Html
+ * @package Boparaiamrit\Datatables\Html
  * @author  Arjay Angeles <aqangeles@gmail.com>
  */
 class Builder
@@ -321,7 +321,7 @@ class Builder
     /**
      * Add a Column object in collection.
      *
-     * @param \Yajra\Datatables\Html\Column $column
+     * @param \Boparaiamrit\Datatables\Html\Column $column
      * @return $this
      */
     public function add(Column $column)