+++ tests/DatatablesTest.php
@@ -1,8 +1,8 @@
 <?php
 
 
-use Yajra\Datatables\Datatables;
-use Yajra\Datatables\Request;
+use Boparaiamrit\Datatables\Datatables;
+use Boparaiamrit\Datatables\Request;
 
 class DatatablesTest extends PHPUnit_Framework_TestCase
 {
@@ -11,7 +11,7 @@ class DatatablesTest extends PHPUnit_Framework_TestCase
         $datatables = $this->getDatatables();
         $html       = $datatables->getHtmlBuilder();
 
-        $this->assertInstanceOf('Yajra\Datatables\Html\Builder', $html);
+        $this->assertInstanceOf('Boparaiamrit\Datatables\Html\Builder', $html);
     }
 
     public function test_get_request()
@@ -19,11 +19,11 @@ class DatatablesTest extends PHPUnit_Framework_TestCase
         $datatables = $this->getDatatables();
         $request    = $datatables->getRequest();
 
-        $this->assertInstanceOf('Yajra\Datatables\Request', $request);
+        $this->assertInstanceOf('Boparaiamrit\Datatables\Request', $request);
     }
 
     /**
-     * @return \Yajra\Datatables\Datatables
+     * @return \Boparaiamrit\Datatables\Datatables
      */
     protected function getDatatables()
     {